+++ src/imio/project/pst/browser/viewlets.py
@@ -8,6 +8,7 @@ from collective.messagesviewlet.browser.messagesviewlet import MessagesViewlet
 from collective.messagesviewlet.message import PseudoMessage
 from imio.helpers.content import richtextval
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
+from plone import api
 from plone.app.layout.viewlets import ViewletBase
 from zope.component import getUtility
 from zope.intid.interfaces import IIntIds
@@ -33,15 +34,21 @@ class TasksListViewlet(OriginalTasksListViewlet):
 
 class ContentLinkViewlet(ViewletBase):
 
-    index = ViewPageTemplateFile('contentlink.pt')
+    index = ViewPageTemplateFile("contentlink.pt")
 
     def content_link(self):
         if self.back_references(self.context):
             return [obj.aq_parent for obj in self.context.back_references()]
         if hasattr(self.context, "_link_portal_type"):
-            ref = [obj for obj in self.context.symbolic_link.to_object.back_references()]
+            ref = [
+                obj for obj in self.context.symbolic_link.to_object.back_references()
+            ]
             ref.append(self.context.symbolic_link.to_object)
-            return [obj.aq_parent for obj in ref if obj.absolute_url() != self.context.absolute_url()]
+            return [
+                obj.aq_parent
+                for obj in ref
+                if obj.absolute_url() != self.context.absolute_url()
+            ]
 
     def back_references(self, context):
         """
@@ -65,13 +72,51 @@ class ContextInformationViewlet(MessagesViewlet):
     """
 
     def getAllMessages(self):
+        ret = []
         if hasattr(self.context, "_link_portal_type"):
 
-            ret = []
-            msg = translate(u"This content is a copy, to modify the original content click on this button ${edit}",
-                            domain='imio.project.pst', context=self.request,
-                            mapping={'edit': '<a href="{0}/edit">Edit</a>'.format(self.context.symbolic_link.to_object.absolute_url())})
-            ret.append(PseudoMessage(msg_type='significant', text=richtextval(msg),
-                       hidden_uid=generate_uid(), can_hide=False))
-            return ret
+            msg = translate(
+                u"This content is a copy, to modify the original content click on this button ${edit}",
+                domain="imio.project.pst",
+                context=self.request,
+                mapping={
+                    "edit": '<a href="{0}/edit">Edit</a>'.format(
+                        self.context.symbolic_link.to_object.absolute_url()
+                    )
+                },
+            )
+            ret.append(
+                PseudoMessage(
+                    msg_type="significant",
+                    text=richtextval(msg),
+                    hidden_uid=generate_uid(),
+                    can_hide=False,
+                )
+            )
 
+        if self.context.portal_type == "operationalobjective":
+            if self.context.planned_end_date:
+                act_planned_end_date = [
+                    act.planned_end_date
+                    for act in api.content.find(
+                        context=self.context,
+                        portal_type=["pstaction", "action_link", "pstsubaction"],
+                    )
+                ]
+                if act_planned_end_date:
+                    if max(act_planned_end_date) > self.context.planned_end_date:
+                        msg = translate(
+                            u"The planned end date of any one of the actions is greater than the planned end date of the operational objective",
+                            domain="imio.project.pst",
+                            context=self.request,
+                        )
+                        ret.append(
+                            PseudoMessage(
+                                msg_type="significant",
+                                text=richtextval(msg),
+                                hidden_uid=generate_uid(),
+                                can_hide=False,
+                            )
+                        )
+
+        return ret